pyproject.toml
[project]
name = "new-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "beanie>=2.0.1",
    "fastapi[standard]>=0.135.1",
    "motor>=3.7.1",
    "pika>=1.3.2",
    "pyjwt[crypto]>=2.12.1",
    "python-mediator>=0.7.0",
]
[tool.fastapi]
entrypoint = "src.app:app"
